requirements.txt
langchain_community==0.2.5
langchain_core==0.2.7
langchain_openai==0.1.8
python-dotenv==1.0.1
streamlit==1.34.0
